\name{blackmen}
\alias{blackmen}
\docType{data}
\title{
Black Men
}
\description{
the dataset for black men
}
\usage{data("blackmen")}
\format{
  A data frame with 10000 observations on the following 7 variables.
  \describe{
    \item{\code{asecwt}}{weight}
    \item{\code{year}}{a factor indicating period groups with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6}}
    \item{\code{age}}{a factor indicating age groups with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9}}
    \item{\code{labforce}}{labor Force participation rate}
    \item{\code{educ}}{education level}
    \item{\code{educr}}{education level}
    \item{\code{educc}}{education level}
  }
}

\keyword{datasets}
