% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{compute_xcoordinate}
\alias{compute_xcoordinate}
\title{Calculate x coordinate value}
\usage{
compute_xcoordinate(p)
}
\arguments{
\item{p}{Period value.}
}
\value{
The coordinate value for x axis.
}
\description{
Calculate x coordinate value for plotting hexagram in visualizing APC-I
results.
}
