% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{compute_ycoordinate}
\alias{compute_ycoordinate}
\title{Calculate y coordinate value}
\usage{
compute_ycoordinate(p, a)
}
\arguments{
\item{p}{Period value}

\item{a}{Age value}
}
\value{
The coordinate value for y axis.
}
\description{
Calculate y coordinate value for plotting hexagram in visualizing APC-I
results.
}
