\name{cpswomen}
\alias{cpswomen}
\docType{data}
\title{
Women's labor force participation data from the 1990 to 2019 Current Population Survey (CPS)
}
\description{
the dataset for women's labor force participation from the 1990 through 2019 CPS.
}
\usage{data("cpswomen")}
\format{
  A data frame with 1,0000 observations and the following 7 variables.
  \describe{
    \item{\code{asecwt}}{weight}
    \item{\code{year}}{a factor indicating period groups with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6}}
    \item{\code{age}}{a factor indicating age groups with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9}}
    \item{\code{labforce}}{labor Force participation rate}
    \item{\code{educ}}{education level}
    \item{\code{educr}}{education level}
    \item{\code{educc}}{education level}
  }
}

\keyword{datasets}
