\name{simulation}
\alias{simulation}
\docType{data}
\title{
Simulated Dataset
}
\description{
A simulated dataset for APC-I analysis.
}
\usage{data("simulation")}
\format{
  A data frame with 1,0000 observations and the following 3 variables.
  \describe{
    \item{\code{y}}{a numeric}
    \item{\code{age}}{a numeric}
    \item{\code{period}}{a numeric}
  }
}
\keyword{datasets}
