\name{whitemen}
\alias{whitemen}
\docType{data}
\title{
White Men
}
\description{
A dataset for white men.
}
\usage{data("whitemen")}
\format{
  A data frame with 10000 observations on the following 7 variables.
  \describe{
    \item{\code{asecwt}}{weight}
    \item{\code{year}}{a factor indicating period groups}
    \item{\code{age}}{a factor indicating age groups}
    \item{\code{labforce}}{labor Force participation rate}
    \item{\code{educ}}{education level}
    \item{\code{educr}}{education level}
    \item{\code{educc}}{education level}
  }
}

\keyword{datasets}
