\name{whitesample}
\alias{whitesample}
\docType{data}
\title{whitesample}

\description{
A sample dataset
}

\usage{
whitesample
}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
test
}
\source{
CPS
}
\references{
Luo
}
\examples{
data(whitesample)
}
\keyword{datasets}
