\name{women9017}
\alias{women9017}
\docType{data}
\title{women9017}

\description{
A sample dataset
}

\usage{
women9017
}
\format{
  A data frame with 100 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
test
}
\source{
CPS
}
\references{
Luo
}
\examples{
data(women9017)
}
\keyword{datasets}
