\name{women9017}
\alias{women9017}
\docType{data}
\title{women9017}

\description{
A sample dataset
}

\usage{
women9017
}
\format{
  A data frame with 1000 observations on the following 23 variables.
  \describe{
    \item{\code{ac}}{a numeric vector}
    \item{\code{acc}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{cc}}{a numeric vector}
    \item{\code{ccc}}{a numeric vector}
    \item{\code{cohort}}{a numeric vector}
    \item{\code{educ}}{a numeric vector}
    \item{\code{educc}}{a numeric vector}
    \item{\code{educr}}{a numeric vector}
    \item{\code{inlfc}}{a numeric vector}
    \item{\code{labforce}}{a numeric vector}
    \item{\code{lfc}}{a numeric vector}
    \item{\code{marst}}{a numeric vector}
    \item{\code{marstc}}{a numeric vector}
    \item{\code{marstr}}{a numeric vector}
    \item{\code{nc}}{a numeric vector}
    \item{\code{ncc}}{a numeric vector}
    \item{\code{nchild}}{a numeric vector}
    \item{\code{pc}}{a numeric vector}
    \item{\code{pcc}}{a numeric vector}
    \item{\code{wt}}{a numeric vector}
    \item{\code{wtsupp}}{a numeric vector}
    \item{\code{year}}{a numeric vector}
  }
}
\details{
test
}
\source{
CPS
}
\references{
Luo and Hodges (2019)
}

